refactor: rename the feature-flags feature to feature-flag-overrides
The old name collided with the feature-flag system itself (FEATURES,
useHasFeature, features/__core__), so it read as the place flags are
defined rather than the place local overrides are stored and edited. The
new name matches the symbols it already contains — featureFlagOverridesSlice,
selectFeatureFlagOverrides, applyFeatureOverrides.
