+++ apps/web/src/features/feature-flag-overrides/hooks/useFeatureFlagEditorData.ts
@@ -4,7 +4,7 @@ import { useGetChainsConfigV2Query } from '@safe-global/store/gateway'
 import { FEATURES, hasFeature } from '@safe-global/utils/utils/chains'
 import { CONFIG_SERVICE_KEY } from '@/config/constants'
 import { useAppSelector } from '@/store'
-import { selectFeatureFlagOverrides } from '@/features/feature-flags/store'
+import { selectFeatureFlagOverrides } from '@/features/feature-flag-overrides/store'
 import useChainId from '@/hooks/useChainId'
 
 export type FeatureFlagRowData = {
